wip: simpler importer API
